Seção 9 - Dtos e data annotations

+++ Back/src/ProEventos/Controllers/EventosController.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using ProEventos.Data;
-
+using ProEventos.Dtos;
 using SQLitePCL;
 using System;
 using System.Collections.Generic;
@@ -33,7 +33,8 @@ namespace ProEventos.Controllers
            try
            {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento por Id encontrado.");
+                if (eventos == null) return NoContent();
+
                 return Ok(eventos);
            }
            catch(Exception ex)
@@ -47,7 +48,7 @@ namespace ProEventos.Controllers
             try
             {
                 var evento = await _eventoService.GetEventosByIdAsync(id, true);
-                if (evento == null) return NotFound("Nenhum evento encontrado.");
+                if (evento == null) return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -61,7 +62,7 @@ namespace ProEventos.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por tema não encontrados");
+                if (evento == null) return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -70,12 +71,12 @@ namespace ProEventos.Controllers
             }
         }
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento");
+                if (evento == null) return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -84,12 +85,12 @@ namespace ProEventos.Controllers
             }
         }
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
-                var evento = await _eventoService.UpdateEvento(id,model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento");
+                var evento = await _eventoService.UpdateEvento(id, model);
+                if (evento == null) return NoContent();
                 return Ok(evento);
             }
             catch (Exception ex)
@@ -102,9 +103,12 @@ namespace ProEventos.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventosByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ?
                   Ok("Deletado") :
-                  BadRequest("evento não deletado");
+                  throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento.");
                 
             }
             catch (Exception ex)
